Use ENV variable for CDN
Summary:
Before, the CDN url was hardcoded in the different places it was needed.
This resulted in a bug in staging where the CDN for production was used
for image URLs.

This diff fixes the CDN so that it's set correctly in all environments.

Test Plan: Deploy, then verify that classroom image uploading works on staging.

Projects: #roots_of_success

Differential Revision: http://phab.calblueprint.org/D130